% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{as_meas}
\alias{as_meas}
\title{Convert to measurement class}
\usage{
as_meas(x)
}
\arguments{
\item{x}{A numeric vector of measurements.}
}
\value{
The input with class "meas" added.
}
\description{
Add "meas" class to a vector of measurements for better printing.
}
\keyword{internal}
