% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{available_measures}
\alias{available_measures}
\title{List available measurements}
\usage{
available_measures()
}
\value{
Character vector of available measurement names.
}
\description{
Show all measurements that can be computed with \code{measure()}.
}
\details{
Returns names of all scalar-returning \verb{get_*} functions that can be
used with \code{measure()}.
}
\examples{
available_measures()

}
\keyword{internal}
