% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momocs2_datasets.R
\docType{data}
\name{bot}
\alias{bot}
\title{Some bottles}
\format{
A \link{mom_tbl} with:
\describe{
\item{coo}{outline coordinates}
\item{type}{of drink}
\item{fake}{a fake vector}
}
}
\source{
todo
}
\usage{
bot
}
\description{
A dataset containing bottles outlines
}
\keyword{datasets}
