% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bot2}
\alias{bot2}
\title{Some bottles}
\format{A data frame with 53940 rows and 10 variables:
\describe{
\item{coo}{some coords}
\item{type}{of drink}
\item{fake}{a fake vector}
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
bot2
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
