% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{coe_single}
\alias{coe_single}
\title{coe_single constructor for single coefficients lists}
\usage{
coe_single(x)
}
\arguments{
\item{x}{anything that can be turned into a coe_single, typically something that looks like a
two columns}
}
\value{
a \code{coe_single} object
}
\description{
\code{coe} objects are \link[tibble:tibble-package]{tibble} with as many rows as created by morphometric methods.
It is not really intended to be used directly but is useful if you want to extend Momocs.
}
\examples{
tibble::tibble(a=1, b=2) \%>\% coe_single()
}
