% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is_coe_tbl}
\alias{is_coe_tbl}
\alias{is_coe_tbl1}
\alias{coe_tbl}
\title{coe_tbl class}
\usage{
is_coe_tbl(x)

is_coe_tbl1(x)

coe_tbl(x)
}
\arguments{
\item{x}{anything sensible}
}
\description{
coe_tbl class
}
\section{Functions}{
\itemize{
\item \code{is_coe_tbl}: Class tester

\item \code{is_coe_tbl1}: Class tester
}}

