% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{coo_align}
\alias{coo_align}
\title{Align shapes}
\usage{
coo_align(x)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_tbl} or a list of shapes}
}
\value{
\link{coo_single} or \link{coo_tbl} or a list of shapes
}
\description{
Align shape along their longer axis using var-cov matrix and eigen values.
}
\examples{
bot2$coo[[5]] \%>\% coo_align
}
\seealso{
Other coo_modifyers: 
\code{\link{coo_center}()},
\code{\link{coo_scale}()},
\code{\link{coo_trans}()}
}
\concept{coo_modifyers}
