% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{coo_center}
\alias{coo_center}
\alias{coo_centre}
\title{Center shapes}
\usage{
coo_center(x)

coo_centre(x)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{coo_tbl}}
}
\value{
a \link{coo_single}, \link{coo_list} or \link{coo_tbl}
}
\description{
Returns a shape centered on the origin.
}
\examples{

bot2 \%>\% pick(1) \%>\% coo_center \%>\% gg()
bot2 \%>\% coo_center \%>\% pile()

}
\seealso{
Other coo_modifyers: 
\code{\link{coo_align}()},
\code{\link{coo_rotate}()},
\code{\link{coo_sample}()},
\code{\link{coo_scale}()},
\code{\link{coo_template}()},
\code{\link{coo_trans}()}

Other translations: 
\code{\link{coo_trans}()}
}
\concept{coo_modifyers}
\concept{translations}
