% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{coo_close}
\alias{coo_close}
\alias{coo_unclose}
\alias{is_closed}
\alias{is_unclosed}
\title{Close and unclose shapes}
\usage{
coo_close(x, from_col, to_col, ...)

coo_unclose(x, from_col, to_col, ...)

is_closed(x)

is_unclosed(x)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{mom_tbl}}

\item{from_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{to_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{...}{useless here}
}
\value{
a \link{coo_single}, \link{coo_list} or \link{mom_tbl}
}
\description{
A closed shape is a shape with the same first and last coordinates.
}
\section{Functions}{
\itemize{
\item \code{coo_unclose}: Unclose shapes

\item \code{is_closed}: Tests if a coo_single (only) is closed

\item \code{is_unclosed}: Tests if a coo_single (only) is unclosed
}}

\examples{

bot \%>\% pick(1)  # todo
}
\concept{coo_modifyers close}
