% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{coo_rotate}
\alias{coo_rotate}
\alias{coo_rotatecenter}
\title{Rotate shapes}
\usage{
coo_rotate(x, theta = 0)

coo_rotatecenter(x, theta = 0, center = c(0, 0))
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{coo_tbl}}

\item{theta}{\code{numeric} angle to rotate (in radians) and in the trigonometric direction (anti-clockwise). Default to \code{0}.}

\item{center}{\code{numeric} of length 2, sepcifying the \verb{(x; y)} coordinates of the rotation center. Default to \code{c(0, 0)}}
}
\value{
a \link{coo_single}, \link{coo_list} or \link{coo_tbl}
}
\description{
Rotates the coordinates by a \code{theta} angle (in radians) in
the trigonometric direction (anti-clockwise).
\code{coo_rotate} uses the origin, but \code{coo_rotatecenter} allows to specify another center.
}
\examples{
x <- bot2 \%>\% pick(1)
gg(x)

x \%>\% coo_rotate(pi/2) \%>\% draw(col="red")
x \%>\% coo_rotate(degrees_to_radians(-45)) \%>\% draw(col="blue")

bot2 \%>\% coo_rotate(pi) \%>\% pile()

}
\seealso{
Other coo_modifyers: 
\code{\link{coo_align}()},
\code{\link{coo_center}()},
\code{\link{coo_sample}()},
\code{\link{coo_scale}()},
\code{\link{coo_template}()},
\code{\link{coo_trans}()}

Other rotations: 
\code{\link{coo_align}()}
}
\concept{coo_modifyers}
\concept{rotations}
