% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{coo_rotatecenter}
\alias{coo_rotatecenter}
\alias{coo_rotatecenter.default}
\alias{coo_rotatecenter.coo_list}
\alias{coo_rotatecenter.list}
\alias{coo_rotatecenter.mom_tbl}
\title{Rotate shapes and specify center}
\usage{
coo_rotatecenter(x, theta = 0, center = c(0, 0), ...)

\method{coo_rotatecenter}{default}(x, theta = 0, center = c(0, 0), ...)

\method{coo_rotatecenter}{coo_list}(x, theta = 0, center = c(0, 0), ...)

\method{coo_rotatecenter}{list}(x, theta = 0, center = c(0, 0), ...)

\method{coo_rotatecenter}{mom_tbl}(
  x,
  theta = 0,
  center = c(0, 0),
  from_col = coo,
  to_col = coo,
  ...
)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{mom_tbl}}

\item{theta}{\code{numeric} angle to rotate (in radians) and in the trigonometric direction (anti-clockwise). Default to \code{0}.}

\item{center}{\code{numeric} of length 2, sepcifying the \verb{(x; y)} coordinates of the rotation center. Default to \code{c(0, 0)}}

\item{...}{useless here}

\item{from_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{to_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}
}
\value{
a \link{coo_single}, \link{coo_list} or \link{mom_tbl}
}
\description{
Rotates the coordinates by a \code{theta} angle (in radians) in
the trigonometric direction (anti-clockwise) and using \code{center} as coordinates.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: default method

\item \code{coo_list}: coo_list method

\item \code{list}: list method

\item \code{mom_tbl}: mom_tbl method
}}

\examples{
x <- bot \%>\% pick(5) \%>\% coo_center() \%>\% coo_scale()
seq(0, pi, pi/6) \%>\%
  purrr::map(~ x \%>\% coo_rotatecenter(x, theta=.x, center=c(5, 5)) \%>\% coo_single) \%>\%
  mom_tbl() \%>\%
  pile()
}
\seealso{
Other coo_modifyers: 
\code{\link{coo_align}()},
\code{\link{coo_center}()},
\code{\link{coo_rotate}()},
\code{\link{coo_sample}()},
\code{\link{coo_scale}()},
\code{\link{coo_template}()},
\code{\link{coo_trans}()}

Other rotations: 
\code{\link{coo_align}()},
\code{\link{coo_rotate}()}
}
\concept{coo_modifyers}
\concept{rotations}
