% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{coo_sample}
\alias{coo_sample}
\alias{coo_sample_prop}
\alias{coo_interpolate}
\title{Sample shapes}
\usage{
coo_sample(x, n, from_col, to_col, ...)

coo_sample_prop(x, prop, from_col = coo, to_col = coo, ...)

coo_interpolate(x, n, from_col, to_col, ...)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{mom_tbl}}

\item{n}{\code{integer} desired number of coordinates (required)}

\item{from_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{to_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{...}{useless here}

\item{prop}{\code{numeric} desired poportion of sampled coordinates (required).
Below 1 will sample, above 1 will interpolate.}
}
\value{
a \link{coo_single}, \link{coo_list} or \link{mom_tbl}
}
\description{
Changes the number of shape coordinates.
}
\details{
\link{coo_sample} will return coordinates regularly sampled along the curvilinear abscissa.
The last point will be dropped so that the distance (new last - first) roughly equals
all other distances between consecutive points. This is typically useful for outlines.
coo_sample_curve (todo link) will preserve the first and last points.
This is typically useful for curves, hence the name.
\link{coo_interpolate} will upsample the number of points.
All functions have a \verb{_prop} ally, where \code{n} is deduced from the proportion you want to retain.
}
\section{Functions}{
\itemize{
\item \code{coo_sample_prop}: Sample a proportion of coordinates

\item \code{coo_interpolate}: Interpolates shape coordinates
}}

\examples{
x <- bot \%>\% pick(1) \%>\% coo_sample(24)
x \%>\% gg()
x \%>\% coo_interpolate(120) \%>\% gg()

bot$coo[1:2] \%>\% coo_sample(12) \%>\% purrr::map_dbl(nrow)
bot$coo[1:2] \%>\% coo_sample(12) \%>\% coo_interpolate(24) \%>\% purrr::map_dbl(nrow)


}
\seealso{
Other coo_modifyers: 
\code{\link{coo_align}()},
\code{\link{coo_center}()},
\code{\link{coo_flip}},
\code{\link{coo_rev}()},
\code{\link{coo_rotatecenter}()},
\code{\link{coo_rotate}()},
\code{\link{coo_sample_rr}()},
\code{\link{coo_scale}()},
\code{\link{coo_slide}()},
\code{\link{coo_template}()},
\code{\link{coo_trans}()},
\code{\link{coo_trim}()},
\code{\link{coo_up}()}
}
\concept{coo_modifyers}
\concept{samplers}
