% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{coo_shear}
\alias{coo_shear}
\title{Shear shapes}
\usage{
coo_shear(x, x_k, y_k, from_col, to_col, ...)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{mom_tbl}}

\item{x_k, y_k}{\code{numeric} shearing factor over x and y axes
(default to 0, ie no shearing)}

\item{from_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{to_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{...}{useless here}
}
\value{
a \link{coo_single}, \link{coo_list} or \link{mom_tbl}
}
\description{
Returns a sheared shape.
}
\examples{

x <- bot \%>\% pick(1) \%>\% coo_center()
x \%>\% gg()
x \%>\% coo_shear(x_k =  0.25) \%>\% gg()
x \%>\% coo_shear(y_k = -0.5) \%>\% gg()
x \%>\% coo_shear(x_k = 0.25, y_k = -0.5) \%>\% gg()
}
\seealso{
Other coo_modifyers: 
\code{\link{coo_align}()},
\code{\link{coo_center}()},
\code{\link{coo_flip}},
\code{\link{coo_rev}()},
\code{\link{coo_rotatecenter}()},
\code{\link{coo_rotate}()},
\code{\link{coo_sample_rr}()},
\code{\link{coo_sample}()},
\code{\link{coo_scale}()},
\code{\link{coo_slide}()},
\code{\link{coo_split}()},
\code{\link{coo_template}()},
\code{\link{coo_trans}()},
\code{\link{coo_trim}()},
\code{\link{coo_up}()}
}
\concept{coo_modifyers}
\concept{shearings}
