% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{coo_smooth}
\alias{coo_smooth}
\title{Smooth shapes}
\usage{
coo_smooth(x, n, from_col, to_col, ...)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{mom_tbl}}

\item{n}{\code{integer} smoothing iterations}

\item{from_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{to_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{...}{useless here}
}
\value{
a \link{coo_single}, \link{coo_list} or \link{mom_tbl}
}
\description{
Smooth shape coordinates using a moving average
}
\examples{

bot \%>\% pick(1) \%>\% coo_smooth(5) \%>\% gg()

}
\concept{coo_modifyers smooth}
