% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{coo_smooth_curve}
\alias{coo_smooth_curve}
\title{Smooth shapes but preserve tips}
\usage{
coo_smooth_curve(x, n, from_col, to_col, ...)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{mom_tbl}}

\item{n}{\code{integer} smoothing iterations}

\item{from_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{to_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{...}{useless here}
}
\value{
a \link{coo_single}, \link{coo_list} or \link{mom_tbl}
}
\description{
\link{coo_smooth} variant that preserves coordinates of first and alst points.
Typically useful for curves
}
\examples{

bot \%>\% pick(1) #todo


}
\seealso{
Other coo_modifyers smooth: 
\code{\link{coo_smooth}()}
}
\concept{coo_modifyers smooth}
