% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{coo_tbl}
\alias{coo_tbl}
\alias{is_coo_tbl}
\alias{is_coo_tbl1}
\title{coo_tbl class}
\usage{
coo_tbl(x)

is_coo_tbl(x)

is_coo_tbl1(x)
}
\arguments{
\item{x}{anything sensible}
}
\description{
coo_tbl class
}
\section{Functions}{
\itemize{
\item \code{is_coo_tbl}: Class tester

\item \code{is_coo_tbl1}: Class tester
}}

