% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{coo_template}
\alias{coo_template}
\title{XXX shapes}
\usage{
coo_template(x, ...)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{coo_tbl}}

\item{...}{additional parameters}

\item{size}{\code{numeric} the side of the square inscribing the shape}
}
\value{
templated \link{coo_list} or \link{coo_tbl}
}
\description{
Centers shape and scale them so that they are inscribed in a \code{size}-side square.
}
\examples{
bot2 \%>\% pick(1) \%>\% coo_template() \%>\% gg()
bot2 \%>\% coo_template \%>\% pile()
}
\seealso{
Other coo_modifyers: 
\code{\link{coo_align}()},
\code{\link{coo_center}()},
\code{\link{coo_scale}()},
\code{\link{coo_trans}()}
}
\concept{coo_modifyers}
