% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{coo_trans}
\alias{coo_trans}
\title{Translate shapes}
\usage{
coo_trans(x, x_trans = 0, y_trans = 0, ...)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{coo_tbl}}

\item{x_trans, y_trans}{\code{numeric} how much translate on each direction}

\item{...}{useless here}
}
\value{
a \link{coo_single}, \link{coo_list} or \link{coo_tbl}
}
\description{
Returns a shape translate by \code{x} and \code{y}.
}
\examples{
bot2$coo[[1]] \%>\% coo_trans

}
\seealso{
Other coo_modifyers: 
\code{\link{coo_align}()},
\code{\link{coo_center}()},
\code{\link{coo_rotate}()},
\code{\link{coo_sample}()},
\code{\link{coo_scale}()},
\code{\link{coo_template}()}

Other translations: 
\code{\link{coo_center}()}
}
\concept{coo_modifyers}
\concept{translations}
