% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{coo_trim}
\alias{coo_trim}
\alias{coo_trim_head}
\alias{coo_trim_tail}
\title{Trim coordinates from shape}
\usage{
coo_trim(x, n, from_col, to_col, ...)

coo_trim_head(x, n, from_col, to_col, ...)

coo_trim_tail(x, n, from_col, to_col, ...)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{mom_tbl}}

\item{n}{\code{integer} how many coordinates shall we trim}

\item{from_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{to_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{...}{useless here}
}
\value{
a \link{coo_single}, \link{coo_list} or \link{mom_tbl}
}
\description{
\link{coo_trim_head} removes the first \code{n} coordinates from shape,
\link{coo_trim_tail} removes the last \code{n} coordinates,
\link{coo_trim} does both.
}
\section{Functions}{
\itemize{
\item \code{coo_trim_head}: Trims head of shape

\item \code{coo_trim_tail}: Trims tail of shape
}}

\examples{

x <- bot \%>\% pick(1) \%>\% coo_sample(12)

coo_trim(x, 5)
coo_trim_head(x, 5)
coo_trim_tail(x, 5)
}
\seealso{
Other coo_modifyers: 
\code{\link{coo_align}()},
\code{\link{coo_center}()},
\code{\link{coo_flip}},
\code{\link{coo_rev}()},
\code{\link{coo_rotatecenter}()},
\code{\link{coo_rotate}()},
\code{\link{coo_sample_rr}()},
\code{\link{coo_sample}()},
\code{\link{coo_scale}()},
\code{\link{coo_slide}()},
\code{\link{coo_template}()},
\code{\link{coo_trans}()},
\code{\link{coo_up}()}
}
\concept{coo_modifyers}
