% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{coo_up}
\alias{coo_up}
\alias{coo_down}
\alias{coo_left}
\alias{coo_right}
\title{Retains coordinates based on their x/y sign}
\usage{
coo_up(x, from_col, to_col, ...)

coo_down(x, from_col, to_col, ...)

coo_left(x, from_col, to_col, ...)

coo_right(x, from_col, to_col, ...)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{mom_tbl}}

\item{from_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{to_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{...}{useless here}
}
\value{
a \link{coo_single}, \link{coo_list} or \link{mom_tbl}
}
\description{
Useful for centered or aligned shapes.
}
\details{
\itemize{
\item \link{coo_up} retains only coordinates with \code{y >= 0}
\item \link{coo_down} retains only coordinates with \code{y <= 0}
\item \link{coo_left} retains only coordinates with \code{x <= 0}
\item \link{coo_right} retains only coordinates with \code{x >= 0}
}
}
\section{Functions}{
\itemize{
\item \code{coo_up}: filter upper part of a shape

\item \code{coo_down}: filter lower part of a shape

\item \code{coo_left}: filter left part of a shape

\item \code{coo_right}: filter right part of a shape
}}

\examples{

bot \%>\% pick(1) \%>\% coo_center() \%>\% coo_left() \%>\% gg()

}
\seealso{
Other coo_modifyers: 
\code{\link{coo_align}()},
\code{\link{coo_center}()},
\code{\link{coo_flip}},
\code{\link{coo_rev}()},
\code{\link{coo_rotatecenter}()},
\code{\link{coo_rotate}()},
\code{\link{coo_sample_rr}()},
\code{\link{coo_sample}()},
\code{\link{coo_scale}()},
\code{\link{coo_shear}()},
\code{\link{coo_slide}()},
\code{\link{coo_split}()},
\code{\link{coo_template}()},
\code{\link{coo_trans}()},
\code{\link{coo_trim}()}
}
\concept{coo_modifyers}
