% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_graphics.R
\name{decent}
\alias{decent}
\alias{decent_decimals}
\alias{decent_geom}
\alias{decent_alpha}
\alias{decent_size}
\alias{decent_size_path}
\alias{decent_size_point}
\alias{decent_size_text}
\title{Decent default values}
\usage{
decent_decimals(x)

decent_geom(x)

decent_alpha(x)

decent_size(x)

decent_size_path(x)

decent_size_point(x)

decent_size_text(x)
}
\arguments{
\item{x}{\code{integer} a number on which to find a sensible value}
}
\description{
Aims at finding sesisble default values when not specified.
}
\details{
Sizes are expressed in typographic points.
Mostly used internally, but can be overriden.
Will possibly be turned into options at some point.
}
\section{Functions}{
\itemize{
\item \code{decent_decimals}: decimals

\item \code{decent_geom}: geom

\item \code{decent_alpha}: alpha

\item \code{decent_size}: size_geom

\item \code{decent_size_path}: size_path

\item \code{decent_size_point}: size_point

\item \code{decent_size_text}: size_text
}}

\examples{
tibble::tibble(x=10^(1:5)) \%>\%
    dplyr::mutate_all(.funs=list(
        decimals   = decent_decimals,
        alpha      = decent_alpha,
        #geom       = decent_geom,
        size_path  = decent_size_path,
        size_point = decent_size_point,
        size_text  = decent_size_text
        ))
}
