% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_dfourier.R
\name{dfourier}
\alias{dfourier}
\title{Discrete cosine transform}
\usage{
dfourier(x, nb_h, raw = FALSE, from_col, to_col, drop_coo, ...)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{mom_tbl}}

\item{nb_h}{\code{int} nb of harmonics. Default to \code{6} for \code{dfourier}, to all of them for \code{dfourier_i}}

\item{raw}{\code{logical} whether to return raw and full results for \link{dfourier}}

\item{drop_coo}{\code{logical} whether to drop coo column (default to \code{TRUE})}

\item{...}{for generics. Useless here.}

\item{nb_pts}{\code{int} nb of points for the reconstruction}

\item{from_coo, to_col}{column names}

\item{from_coe}{column name for inverse method}
}
\value{
a list with components when applied on a single shape:
}
\description{
Calculates discrete cosine transforms, as introduced by Dommergues and colleagues, on
open outlines.
}
\details{
Shapes must be normalized, ie using \link{coo_bookstein} before performing the dct transform.
}
\examples{
olea[1:3, ] \%>\% dfourier()
}
\references{
\itemize{
\item Dommergues, C. H., Dommergues, J.-L., & Verrecchia, E. P. (2007).
The Discrete Cosine Transform, a Fourier-related Method for Morphometric Analysis of Open Contours.
\emph{Mathematical Geology}, 39(8), 749-763. doi:10.1007/s11004-007-9124-6
\item Many thanks to Remi Laffont for the original translation in R).
}
}
\seealso{
Other morphometrics: 
\code{\link{efourier}()},
\code{\link{npoly}()},
\code{\link{opoly}()}
}
\concept{morphometrics}
