% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_graphics.R
\name{draw}
\alias{draw}
\title{Add shapes on top of another plot}
\usage{
draw(x, gg, ...)
}
\arguments{
\item{x}{a Momocs object}

\item{gg}{\code{ggplot} object, default to \link[ggplot2:last_plot]{ggplot2::last_plot}}

\item{...}{additional parameters to feed geoms}
}
\value{
a \code{ggplot} object
}
\description{
Add shapes on top of another plot
}
\examples{
bot \%>\% pick(1) \%>\% gg()
bot \%>\% pick(2) \%>\% draw()

bot \%>\% pile(alpha=0.2)
bot \%>\% coo_rotate(pi/2) \%>\% draw(col="slateblue")
}
