% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{edi}
\alias{edi}
\title{Distance helpers}
\usage{
edi(x, y, r = 0.5)
}
\arguments{
\item{x, y}{\link{coo_single}}

\item{r}{\code{numeric} how much of the distance \code{d(x -> y)} should we travel?}
}
\description{
A small tribe to help with euclidean distances calculations
}
