% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-g.R
\name{g}
\alias{g}
\title{ggplot2 plotter}
\usage{
g(x, ...)
}
\arguments{
\item{x}{object with a defined method}

\item{...}{additional parameters}
}
\description{
Prepare and build ggplot2 graphics from common classes
}
