% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_descriptors.R
\name{get_centpos}
\alias{get_centpos}
\title{Calculate centroid position}
\usage{
get_centpos(x)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_tbl} or a list of shapes}
}
\value{
\code{numeric} or additional columns
}
\description{
Simply the average of \code{x} and \code{y}.
}
\details{
This function can be used to integrate size - if meaningful -
}
\examples{

bot2$coo[[1]] \%>\% get_centpos

}
\seealso{
Other coo_descriptors: 
\code{\link{get_centsize}()}
}
\concept{coo_descriptors}
