% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldk_handlers.R
\name{get_ldk}
\alias{get_ldk}
\title{Extract landmark coordinates}
\usage{
get_ldk(x, from_col = coo, ldk_col = ldk)
}
\arguments{
\item{x}{\link{mom_tbl}}

\item{from_col}{column name (\code{coo} by default)}

\item{ldk_col}{column name (\code{ldk} by default)}
}
\description{
From a \link{mom_tbl} with coordinates (eg outlines or curves) \emph{and}
a landmark column (a list of row ids being landmarks for each shape).
}
\examples{
h <- hearts \%>\% get_ldk()
h$coo_ldk \%>\% head()
}
\concept{ldk with a new column named {{from_col}}_{{ldk_col}}.}
