% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_getters.R
\name{get_lw}
\alias{get_lw}
\alias{get_length}
\alias{get_width}
\title{Calculate length and width of a shape}
\usage{
get_lw(x)

get_length(x)

get_width(x)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{mom_tbl}}
}
\value{
\code{numeric} or additional column
}
\description{
Returns the length and width of a shape. The shape is first aligned using
\link{coo_align}, then length is the range along the x-axis and the width as
the range on the y-axis.
}
\examples{
bot$coo[[5]] \%>\% get_lw
}
\concept{coo_ descriptors}
