% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_getters.R
\name{get_nb}
\alias{get_nb}
\title{Calculate the number of coordinates per shape}
\usage{
get_nb(x)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{coo_tbl}}
}
\value{
\code{numeric} or additional column
}
\description{
Wraps around \code{nrow}
}
\examples{
bot2 \%>\% pick(1) \%>\% get_nb()
bot2 \%>\% get_nb()
}
\seealso{
Other getters: 
\code{\link{get_perim_along}()}
}
\concept{getters}
\concept{shape getters}
