% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_getters.R
\name{get_range}
\alias{get_range}
\title{Get shape range}
\usage{
get_range(x)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_tbl} or a list of shapes}
}
\value{
\code{numeric} or additional column
}
\description{
Just a wrapper around \link{range}
}
\examples{
bot2$coo[[5]] \%>\% get_range

}
\seealso{
Other coo_descriptors: 
\code{\link{get_centpos}()},
\code{\link{get_centsize}()}
}
\concept{coo_descriptors}
