% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_graphics.R
\name{gg}
\alias{gg}
\alias{gg0}
\title{Default ggplot2 graphics}
\usage{
gg(x, first = TRUE, centroid = TRUE, axes = TRUE, ...)

gg0(x, ...)
}
\arguments{
\item{x}{a Momocs object}

\item{first}{\code{logical} whether to draw first point}

\item{centroid}{\code{logical} whether to draw centroid}

\item{axes}{\code{logical} whether to draw axes, text and grid}

\item{...}{additional parameters to feed geoms}

\item{gg}{\code{ggplot} object, default to \link[ggplot2:last_plot]{ggplot2::last_plot}}
}
\value{
a \code{ggplot} object
}
\description{
Default (ggplot2) visualisations for Momocs objects.
}
\details{
\code{gg} is the base plotter.
\code{gg0} prepare the canvas but let you pick your \code{ggplot2::geoms}.
\code{draw} add shapes on top of last plot
}
\examples{
bot2 \%>\% pick(1) \%>\% gg()
bot2 \%>\% pick(1) \%>\% gg0() + ggplot2::geom_point(shape="circle plus")
}
