% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momocs2_datasets.R
\docType{data}
\name{hearts}
\alias{hearts}
\title{Outlines of hearts}
\format{
An object of class \code{mom_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 240 rows and 3 columns.
}
\source{
We thank the fellows of the Ecology Department of
the French Institute of Pondicherry that drawn the hearts.
They then have been scanned, smoothed, scaled, centered,
and downsampled to 80 coordinates per outline.
}
\usage{
hearts
}
\description{
A dataset containing 240 hand-drawn hearts with 4 landmarks
}
\keyword{datasets}
