% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_graphics.R
\name{inspect}
\alias{inspect}
\title{Graphical inspection of shapes}
\usage{
inspect(x, ...)
}
\arguments{
\item{x}{a Momocs object}

\item{...}{additional parameters forwarded to \link{gg}}
}
\description{
Sample one shape, plot it with \link{gg}, repeat.
}
\examples{
\dontrun{
bot \%>\% inspect
}

}
\seealso{
Other family_picture: 
\code{\link{mosaic}()}
}
\concept{family_picture}
