% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momocs_datasets.R
\docType{data}
\name{jb}
\alias{jb}
\title{Outline of jb bottle}
\format{
A \link{coo_single} with:
\describe{
\item{x}{coordinates}
\item{y}{coordinates}
}
}
\source{
todo
}
\usage{
jb
}
\description{
A \link{coo_single} obtained with \code{pick(bot, "jb")}
}
\keyword{datasets}
