% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom_coo_coe_classes.R
\name{mom_tbl}
\alias{mom_tbl}
\alias{mom}
\alias{new_mom}
\alias{is_mom_tbl}
\alias{is_mom_tbl1}
\title{Declare MomX tibble}
\usage{
new_mom(x = tibble())

mom(x)

is_mom_tbl(x)

is_mom_tbl1(x)
}
\arguments{
\item{x}{a tibble that may contain column lists}
}
\description{
The main class used accros MomX ecosystem. Basically, a tibble with benefits.
}
\section{Functions}{
\itemize{
\item \code{new_mom}: Constructor

\item \code{mom}: Helpers

\item \code{is_mom_tbl}: Class tester

\item \code{is_mom_tbl1}: Class1 tester
}}

