% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momocs2_package.R
\docType{package}
\name{Momocs2-package}
\alias{Momocs2}
\alias{Momocs2-package}
\title{Momocs2: What the Package Does (One Line, Title Case)}
\description{
To learn more about Momocs2, start with the vignettes:
\code{browseVignettes(package = "Momocs2")}
}
\author{
\strong{Maintainer}: Vincent Bonhomme \email{bonhomme.vincent@gmail.com} (\href{https://orcid.org/0000-0002-2742-6349}{ORCID})

Authors:
\itemize{
  \item Julien Claude (core functions in base R)
}

}
