% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphospaces.R
\name{morphospace}
\alias{morphospace}
\title{Versatile function to reconstruct shapes from statistical objects}
\usage{
morphospace(x, xy)
}
\arguments{
\item{x}{a statistical object. Currently supported: stat_pca}

\item{xy}{a tibble with positions to calculate. Default to morphospace_range on PC1 and PC2}
}
\value{
\link{coo_tbl} with raw shapes and other useful columns
}
\description{
Versatile function to reconstruct shapes from statistical objects
}
\examples{
x <- bot2 \%>\% efourier(6) \%>\% stat_pca()
pos1 <- tibble::tibble(PC1=seq(-5, 5, 0.1))
pos2 <- tibble::tibble(PC1=seq(-5, 5, 0.1), PC2=seq(-5, 5, 0.1))
m <- x \%>\% morphospace(pos1)
m \%>\% pile()
m \%>\% morphospace_template()
}
