% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphospaces.R
\name{morphospace_template}
\alias{morphospace_template}
\title{Template and adjust for multiviews before drawing morphospaces}
\usage{
morphospace_template(x, size = NULL)
}
\arguments{
\item{x}{\link{coo_tbl} returned by \link{morphospace}}

\item{size}{\code{numeric} or \code{ggplot} object.
To fine tune shape size or get a decent adjustment, respectively.}
}
\value{
\link{coo_tbl} ready for ggplot.
}
\description{
Digest information contained in the \link{coo_tbl} returned by \link{morphospace},
by adjusting for position (with \link{coo_trans}), overall templating (\link{coo_template}) and
multiviews adjustment (based on \code{x$name}).
}
\examples{
p <- bot2 \%>\% efourier(6) \%>\% stat_pca()
p \%>\% morphospace() \%>\% pile()
p \%>\% morphospace() \%>\% morphospace_template() \%>\% pile()

#gg <- p \%>\% gg0()
#shps <- p \%>\% morphospace() \%>\% morphospace_template(gg)
}
