% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_graphics.R
\name{mosaic}
\alias{mosaic}
\title{Plot all shapes as a mosaic}
\usage{
mosaic(x, ...)
}
\arguments{
\item{x}{a Momocs object}

\item{...}{additional parameters to feed \code{geom(...)}.}

\item{f}{a factor column to skip lines after the end of each level}

\item{geom}{which \code{ggplot2::geom} to use (\link[ggplot2:geom_polygon]{ggplot2::geom_polygon} by default)}

\item{ncol}{\code{integer} that specifies the number of columns. Otherwise, a roughly squared mosaic is produced.}
}
\value{
a \code{ggplot} object
}
\description{
Plot all shapes as a mosaic
}
\examples{
bot2 \%>\% mosaic()
bot2 \%>\% mosaic(type, ncol=6)
mosaic(bot2, type, ncol=6, aes(col=type))

}
\seealso{
Other family_picture: 
\code{\link{inspect}()},
\code{\link{pile}()}
}
\concept{family_picture}
