% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momocs2_datasets.R
\docType{data}
\name{olea}
\alias{olea}
\title{Outlines of olea stones}
\format{
A \link{mom_tbl} with:
\describe{
\item{coo}{curve coordinates}
\item{var}{olive taxa}
\item{domes}{domestication status}
\item{view}{picture view}
\item{ind}{individual number}
}
}
\source{
todo
}
\usage{
olea
}
\description{
Outlines of olea stones
}
\keyword{datasets}
