% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-g.R
\name{pile}
\alias{pile}
\alias{pile_column}
\title{Plots all shapes on the shape graph}
\usage{
pile(x, ...)

pile_column(x)
}
\arguments{
\item{x}{a coo_tbl object}

\item{...}{additional parameters to feed ggplot2::geoms}
}
\description{
Plots all shapes on the shape graph
}
\details{
\code{pile_column} is an helper function that takes a \code{coo_list} and
returns a single \code{tbl} with an additional name/index column.
}
\note{
former stack in old Momocs
}
\examples{

bot2$coo \%>\% pile_column()
bot2 \%>\% pile()
}
\seealso{
Other family_picture: 
\code{\link{inspect}()}
}
\concept{family_picture}
