% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momocs_verbs.R
\name{plint}
\alias{plint}
\title{Plot, print and pipe forward}
\usage{
plint(x, ...)
}
\arguments{
\item{x}{object to plint}

\item{...}{additional parameters}
}
\value{
what was passed to it (\code{x}), untouched. Side effects are plotting and printing
}
\description{
Within a pipe, will take left hand side and:
i)  produce the default graphics (using \link{gg}),
ii) print the object  (using \link{print}) and
iii) forward in untouched to right hand side of the pipe.
}
