% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom_coo_coe_verbs.R
\name{press}
\alias{press}
\title{Flatten coe list columns}
\usage{
press(x, keep_others)
}
\arguments{
\item{x}{\link{mom_tbl}}

\item{keep_others}{logical whether to return other columns (\code{TRUE} and default`) or only coe}
}
\value{
\link[tibble:tibble]{tibble::tibble}
}
\description{
Take all coe list columns, prefix them with list column names and return a tidy tibble.
This function is typically used for analyses, after morphometrics.
}
\examples{
iris \%>\% mom \%>\% press

x <- bot \%>\% efourier(4)
x$coe2 <- bot$coo \%>\% efourier(6)
x \%>\% press

x \%>\% press(keep_others=FALSE) # eg for PCA

}
\seealso{
\link{unfold} for a related reshaping

Other mom verbs: 
\code{\link{pick}()},
\code{\link{slive}()},
\code{\link{unfold}()}
}
\concept{mom verbs}
