% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{rinse}
\alias{rinse}
\title{Clear the environment}
\usage{
rinse(env = .GlobalEnv)
}
\arguments{
\item{env}{Environment to clear. Default is the global environment.}
}
\value{
Invisibly returns NULL.
}
\description{
Remove all objects from the current environment.
}
\details{
Removes all user-created objects from the specified environment.
Useful for cleaning up workspace between analyses or starting fresh.

Use with caution: removed objects cannot be recovered.
}
\note{
I know J Bryan and my old friend T Poisot said it was anti-social in
a famous tweet but I still find it useful.
}
\examples{
\dontrun{
# Create some objects
x <- 1
y <- 2
z <- data.frame(a = 1:10)

# Check what's in environment
ls()

# Clean it all up
rinse()

# Environment is now empty
ls()
}

}
\keyword{internal}
