% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_pca.R
\name{scree}
\alias{scree}
\alias{scree.default}
\alias{scree.pca}
\alias{scree_min}
\alias{scree_plot}
\title{How many axes to retain this much of variance or trace ?}
\usage{
scree(x, n_axes)

\method{scree}{default}(x, n_axes)

\method{scree}{pca}(x, n_axes)

scree_min(x, prop)

scree_plot(x, n_axes, prop)
}
\arguments{
\item{x}{a stat object (\code{scree}) or a tibble returned by \code{scree} for \verb{scree_*}}

\item{n_axes}{\code{numeric} range of axes to consider.
All by default for \code{scree_min}, display until \code{0.99} for \code{scree_plot}}

\item{prop}{numeric how many axes are enough to gather this proportion of variance.
Default to 1, all axes are returned
defaut to 1: all axis are returned}
}
\value{
scree returns a data.frame, scree_min a numeric, scree_plot a ggplot.
}
\description{
A set of functions around PCA/LDA eigen/trace.
}
\details{
\itemize{
\item \code{scree} calculates their proportion and cumulated proportion;
\item \code{scree_min} returns the minimal number of axis to use to retain a given proportion;
\item \code{scree_plot} displays a screeplot.
}
}
\section{Functions}{
\itemize{
\item \code{scree_min}: minimal number of axes to retain

\item \code{scree_plot}: minimal number of axes to retain
}}

\examples{
p <- bot2 \%>\% efourier(6) \%>\% stat_pca()
p \%>\% scree()
p \%>\% scree_min(0.99)
p \%>\% scree_plot(prop=0.95)
}
