% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shapes}
\alias{shapes}
\title{Shapes dataset}
\format{
a raw list with 4 shapes
}
\source{
Borrowed default shapes from (c) Adobe Photoshop. Do not send me to jail.
}
\usage{
shapes
}
\description{
A list of 4 shapes (as outlines)
}
\keyword{datasets}
