% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom_coo_coe_verbs.R
\name{slive}
\alias{slive}
\title{Pick a single row from a tbl}
\usage{
slive(x, i)
}
\arguments{
\item{x}{a \code{mom_tbl} (or a \link[tibble:tibble]{tibble::tibble} or a \link{data.frame})}

\item{i}{\code{int} which to extract. If not provided, pick one randomly.}
}
\value{
a \link{mom_tbl} (or a \link[tibble:tibble]{tibble::tibble} or a \link{data.frame})
}
\description{
Just a wrapper around \code{[} (so a raw \link[dplyr:slice]{dplyr::slice} that also add the random aspect.
}
\examples{
bot \%>\% slive(1)
}
\seealso{
\link{pick}

Other mom verbs: 
\code{\link{pick}()},
\code{\link{press}()},
\code{\link{unfold}()}
}
\concept{mom verbs}
