% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_pca.R
\name{stat_pca}
\alias{stat_pca}
\title{Principal component analysis}
\usage{
stat_pca(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{\link{coe_tbl}}

\item{center}{\code{logical} whether the variables should be shifted to
be zero centered (\code{center} in \link[stats:prcomp]{stats::prcomp})}

\item{scale}{\code{logical} whether the variables should be scaled to
have unit variance before the analysis takes place (\code{scale.} in \link[stats:prcomp]{stats::prcomp})}
}
\value{
object of class \code{pca}, a \link[tibble:tibble]{tibble::tibble} with attributes
}
\description{
Performs a principal components analysis (delegated to \link[stats:prcomp]{stats::prcomp}).
}
\examples{
(x <-  bot \%>\% efourier(4) \%>\% stat_pca())
attributes(x)
}
\references{
\link[stats:prcomp]{stats::prcomp}
}
