% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momocs2_tidyeval.R
\name{tidyeval_helpers}
\alias{tidyeval_helpers}
\alias{tidyeval_coo_modifyers}
\alias{tidyeval_coo_and_ldk}
\title{Tidyeval helpers}
\usage{
tidyeval_coo_modifyers(from_col, to_col)

tidyeval_coo_and_ldk(from_col, ldk_col)
}
\arguments{
\item{from_col}{column name}

\item{to_col}{column name}

\item{ldk_col}{column name}
}
\value{
\link{closure}, for delayer evaluation
}
\description{
Pack repeated tidyeval behaviour.
Expose to ease development (todo extending vignettes, todo everything here)
}
\details{
\itemize{
\item tidyeval_coo_modifyers: used in \verb{coo_*} modifyers.
\itemize{
\item Default to \code{coo} for both \code{from_col} and \code{to_col}.
\item If \code{from_col} is passed but not \code{to_col}, the latter also uses  \code{from_col}.
This is where the \emph{modifying} behaviour is reflected.
}
}
}
\section{Functions}{
\itemize{
\item \code{tidyeval_coo_modifyers}: coo_ helpers

\item \code{tidyeval_coo_and_ldk}: ldk helpers
}}

\examples{
#tidyeval_coo_modifyers()
#tidyeval_coo_modifyers(plip)
#tidyeval_coo_modifyers(to_col=plop)
#tidyeval_coo_modifyers(plip, plop)

}
