% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wings}
\alias{wings}
\title{Wings dataset}
\format{
A data frame with 20 rows and two columns
\describe{
\item{coo}{18 landmark coordinates}
\item{group}{species}
}
}
\source{
Rohlf and Slice 1990.
}
\usage{
wings
}
\description{
An landmark toy dataset of mosquito of mosquito wings. Modified from Rohlf and Slice 1990.
}
\keyword{datasets}
