% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babel.R
\name{babel}
\alias{babel}
\alias{complex_2_cartesian}
\alias{cartesian_2_complex}
\title{Convert between numeric and classes systems}
\usage{
complex_2_cartesian(x)

cartesian_2_complex(x)
}
\arguments{
\item{x}{some object to convert}
}
\value{
converted object
}
\description{
Function names should be explicit
}
\section{Functions}{
\itemize{
\item \code{complex_2_cartesian}: complex numbers

\item \code{cartesian_2_complex}: complex numbers
}}

\examples{
bot2 \%>\% pick(1) \%>\% head() \%>\%
  cartesian_2_complex() \%T>\% print() \%>\%
  complex_2_cartesian()
}
