% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bot}
\alias{bot}
\title{Bottles dataset}
\format{
A data frame with 20 rows and three columns
\describe{
\item{coo}{outlines coordinates}
\item{type}{of beverage}
\item{dummy}{one more column to play with}
}
}
\source{
Images have been grabbed on the internet and
prepared by the package's authors. No particular choice has been
made on the dimension of the original images or the brands cited here.
An original idea that came up drinking chai in the courtyard of the
French Institute of Pondicherry circa 2011 an that eventually led to Momocs.
}
\usage{
bot
}
\description{
An outline toy dataset with the historical Momocs bottles dataset
}
\keyword{datasets}
