% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom_coo_coe_classes.R
\name{coe_single}
\alias{coe_single}
\alias{is_coe_single}
\alias{is_coe_single1}
\title{Create single coefficients}
\usage{
is_coe_single(x)

is_coe_single1(x)
}
\arguments{
\item{x}{anything that can be turned into a coe_single, typically something that looks like a
two columns}
}
\value{
a \code{coe_single} object
}
\description{
\code{coe} objects are \link[tibble:tibble-package]{tibble} with as many rows as created by morphometric methods.
It is not really intended to be used directly but is useful if you want to extend Momocs.
}
\section{Functions}{
\itemize{
\item \code{is_coe_single}: Class tester

\item \code{is_coe_single1}: Class1 tester
}}

\examples{
matrix(1:12, nrow=1) \%>\% coe_single()
}
