% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{coo_align}
\alias{coo_align}
\alias{coo_align_xax}
\title{Align shapes}
\usage{
coo_align(x, from_col = coo, to_col = coo, ...)

coo_align_xax(x, from_col, to_col, ...)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{mom_tbl}}

\item{from_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{to_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{...}{useless here}
}
\value{
a \link{coo_single}, \link{coo_list} or \link{mom_tbl}
}
\description{
Align shape along their longer axis using var-cov matrix and eigen values.
}
\details{
(todo) For \code{coo_align_xax}: ff some shapes are upside-down
(or mirror of each others), try redefining a new starting point (eg with coo_slidedirection) before
the alignment step. This may solve your problem because coo_calliper orders the `$arr.ind`` used by
coo_aligncalliper.
}
\section{Functions}{
\itemize{
\item \code{coo_align_xax}: align the longest axis of the shape along the x-axis
}}

\examples{
bot \%>\% pick(1) \%>\% coo_align
bot \%>\% pick(1) \%>\% coo_align_xax
}
\seealso{
Other coo_modifyers: 
\code{\link{coo_center}()},
\code{\link{coo_rotatecenter}()},
\code{\link{coo_rotate}()},
\code{\link{coo_sample_rr}()},
\code{\link{coo_sample}()},
\code{\link{coo_scale}()},
\code{\link{coo_template}()},
\code{\link{coo_trans}()},
\code{\link{coo_up}()}

Other rotations: 
\code{\link{coo_rotatecenter}()},
\code{\link{coo_rotate}()}
}
\concept{coo_modifyers}
\concept{rotations}
