% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{coo_align}
\alias{coo_align}
\title{Align shapes}
\usage{
coo_align(x, ...)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{coo_tbl}}

\item{...}{useless here}
}
\value{
a \link{coo_single}, \link{coo_list} or \link{coo_tbl}
}
\description{
Align shape along their longer axis using var-cov matrix and eigen values.
}
\examples{
bot2$coo[[5]] \%>\% coo_align
}
\seealso{
Other coo_modifyers: 
\code{\link{coo_center}()},
\code{\link{coo_rotate}()},
\code{\link{coo_sample}()},
\code{\link{coo_scale}()},
\code{\link{coo_template}()},
\code{\link{coo_trans}()}

Other rotations: 
\code{\link{coo_rotate}()}
}
\concept{coo_modifyers}
\concept{rotations}
