% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{coo_center}
\alias{coo_center}
\alias{coo_centre}
\title{Center shapes}
\usage{
coo_center(x)

coo_centre(x)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_tbl} or a list of shapes}
}
\value{
\link{coo_single} or \link{coo_tbl} or a list of shapes
}
\description{
Returns a shape centered on the origin.
}
\examples{

bot2$coo[[1]] \%>\% coo_center

}
\seealso{
Other coo_modifyers: 
\code{\link{coo_align}()},
\code{\link{coo_scale}()},
\code{\link{coo_trans}()}
}
\concept{coo_modifyers}
