% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{coo_center}
\alias{coo_center}
\alias{coo_centre}
\alias{coo_center.coo_single}
\alias{coo_center.coo_list}
\alias{coo_center.mom_tbl}
\title{Center shapes}
\usage{
coo_center(x, from_col = coo, to_col = coo, ...)

\method{coo_center}{coo_single}(x, ...)

\method{coo_center}{coo_list}(x, ...)

\method{coo_center}{mom_tbl}(x, from_col = coo, to_col = {
    {
        from_col
    }
}, ...)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{mom_tbl}}

\item{from_col, to_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{...}{useless here}
}
\value{
a \link{coo_single}, \link{coo_list} or \link{mom_tbl}
}
\description{
Returns a shape centered on the origin.
}
\section{Methods (by class)}{
\itemize{
\item \code{coo_single}: coo_single method

\item \code{coo_list}: list method

\item \code{mom_tbl}: mom_tbl method
}}

\examples{

bot \%>\% pick(1) \%>\% coo_center \%>\% gg()
bot \%>\% coo_center \%>\% pile()

}
\seealso{
Other coo_modifyers: 
\code{\link{coo_align}()},
\code{\link{coo_flip}},
\code{\link{coo_rev}()},
\code{\link{coo_rotatecenter}()},
\code{\link{coo_rotate}()},
\code{\link{coo_sample_rr}()},
\code{\link{coo_sample}()},
\code{\link{coo_scale}()},
\code{\link{coo_slide}()},
\code{\link{coo_template}()},
\code{\link{coo_trans}()},
\code{\link{coo_trim}()},
\code{\link{coo_up}()}

Other translations: 
\code{\link{coo_trans}()}
}
\concept{coo_modifyers}
\concept{translations}
