% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{coo_list}
\alias{coo_list}
\alias{is_coo_list}
\alias{is_coo_list1}
\title{coo_list constructor for list of shapes}
\usage{
coo_list(x)

is_coo_list(x)

is_coo_list1(x)
}
\arguments{
\item{x}{anything that can be turned into a \code{coo_list}, typically a list of \link{coo_single}}
}
\value{
a \code{coo_list} object
}
\description{
\code{coo_list} objects are \link{list} of \link{coo_single}.
}
\details{
They behave like lists
and thus as regular \link{tibble}/\link{data.frame} columns.
The only difference is that and are understood by morphometric methods used in MomX.
Front users are not likely to use them directly.
}
\section{Functions}{
\itemize{
\item \code{is_coo_list}: Class tester

\item \code{is_coo_list1}: Class tester
}}

\examples{
matrix(1:12, ncol=2) \%>\% list() \%>\% coo_list()
}
