% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom_coo_coe_classes.R
\name{coo_list}
\alias{coo_list}
\alias{new_coo_list}
\alias{vec_ptype_full.coo_list}
\alias{vec_ptype_abbr.coo_list}
\alias{is_coo_list}
\alias{is_coo_list1}
\title{Create list of coo_single}
\usage{
new_coo_list(x = list())

coo_list(x)

vec_ptype_full.coo_list(x)

vec_ptype_abbr.coo_list(x)

is_coo_list(x)

is_coo_list1(x)
}
\arguments{
\item{x}{a list of \link{coo_single}}
}
\value{
a \code{coo_list} object
}
\description{
\code{coo_list} objects are \link{list} of \link{coo_single}.
}
\details{
They behave like lists
and thus as regular \link{tibble}/\link{data.frame} columns.
The only difference is that and are understood by morphometric methods used in MomX.
}
\section{Functions}{
\itemize{
\item \code{new_coo_list}: Constructor

\item \code{coo_list}: Helper

\item \code{is_coo_list}: Class tester

\item \code{is_coo_list1}: Class1 tester
}}

\examples{
matrix(1:12, ncol=2) \%>\% coo_single() \%>\% list() \%>\% coo_list()
}
