% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{degrees_to_radians}
\alias{degrees_to_radians}
\alias{radians_to_degrees}
\alias{coo_rotate}
\alias{coo_rotatecenter}
\title{Rotate shapes}
\usage{
degrees_to_radians(x)

radians_to_degrees(x)

coo_rotate(x, theta = 0)

coo_rotatecenter(x, theta, center = c(0, 0))
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_tbl} or a list of shapes}

\item{theta}{\code{numeric} angle to rotate (in radians) and in the trigonometric direction (anti-clockwise)}

\item{center}{\code{numeric} of length 2, sepcifying the \verb{(x; y)} coordinates of the rotation center}
}
\value{
a \link{coo_single}, \link{coo_list} or \link{coo_tbl}
}
\description{
Rotates the coordinates by a \code{theta} angle (in radians) in
the trigonometric direction (anti-clockwise).
\code{coo_rotate} uses the origin, but \code{coo_rotatecenter} allows to specify another center.
\code{degrees_to_radians} and \code{radians_to_degrees} helps convert between systems.
}
\examples{
x <- bot2 \%>\% pick(1)
gg(x)

x \%>\% coo_rotate(pi/2) \%>\% draw(col="red")
x \%>\% coo_rotate(degrees_to_radians(-45)) \%>\% draw(col="blue")

bot2 \%>\% coo_rotate(pi) \%>\% pile()

}
