% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{coo_rotate}
\alias{coo_rotate}
\alias{coo_rotate.mom_tbl}
\title{Rotate shapes}
\usage{
coo_rotate(x, theta = 0, from_col = coo, to_col = coo, ...)

\method{coo_rotate}{mom_tbl}(
  x,
  theta = 0,
  from_col = coo,
  to_col = {     {         from_col     } },
  ...
)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{mom_tbl}}

\item{theta}{\code{numeric} angle to rotate (in radians) and in the trigonometric direction (anti-clockwise). Default to \code{0}.}

\item{from_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{to_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{...}{useless here}
}
\value{
a \link{coo_single}, \link{coo_list} or \link{mom_tbl}
}
\description{
Rotates the coordinates by a \code{theta} angle (in radians) in
the trigonometric direction (anti-clockwise).
}
\section{Methods (by class)}{
\itemize{
\item \code{mom_tbl}: mom_tbl method
}}

\examples{
x <- bot \%>\% pick(1)
gg(x)

x \%>\% coo_rotate(pi/2) \%>\% draw(col="red")
x \%>\% coo_rotate(degrees_to_radians(-45)) \%>\% draw(col="blue")

bot \%>\% coo_rotate(pi) \%>\% pile()

}
\seealso{
Other coo_modifyers: 
\code{\link{coo_align}()},
\code{\link{coo_center}()},
\code{\link{coo_flip}},
\code{\link{coo_rev}()},
\code{\link{coo_rotatecenter}()},
\code{\link{coo_sample_rr}()},
\code{\link{coo_sample}()},
\code{\link{coo_scale}()},
\code{\link{coo_slide}()},
\code{\link{coo_template}()},
\code{\link{coo_trans}()},
\code{\link{coo_trim}()},
\code{\link{coo_up}()}

Other rotations: 
\code{\link{coo_align}()},
\code{\link{coo_rotatecenter}()}
}
\concept{coo_modifyers}
\concept{rotations}
