% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{coo_rotatecenter}
\alias{coo_rotatecenter}
\title{Rotate shapes and specify center}
\usage{
coo_rotatecenter(
  x,
  theta = 0,
  center = c(0, 0),
  from_col = coo,
  to_col = coo,
  ...
)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{mom_tbl}}

\item{theta}{\code{numeric} angle to rotate (in radians) and in the trigonometric direction (anti-clockwise). Default to \code{0}.}

\item{center}{\code{numeric} of length 2, sepcifying the \verb{(x; y)} coordinates of the rotation center. Default to \code{c(0, 0)}}

\item{from_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{to_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{...}{useless here}
}
\value{
a \link{coo_single}, \link{coo_list} or \link{mom_tbl}
}
\description{
Rotates the coordinates by a \code{theta} angle (in radians) in
the trigonometric direction (anti-clockwise) and using \code{center} as coordinates.
}
\examples{
x <- bot \%>\% pick(5) \%>\% coo_center() \%>\% coo_scale()
}
\seealso{
Other coo_modifyers: 
\code{\link{coo_align}()},
\code{\link{coo_center}()},
\code{\link{coo_flip}},
\code{\link{coo_rev}()},
\code{\link{coo_rotate}()},
\code{\link{coo_sample_rr}()},
\code{\link{coo_sample}()},
\code{\link{coo_scale}()},
\code{\link{coo_slide}()},
\code{\link{coo_split}()},
\code{\link{coo_template}()},
\code{\link{coo_trans}()},
\code{\link{coo_trim}()},
\code{\link{coo_up}()}

Other rotations: 
\code{\link{coo_align}()},
\code{\link{coo_rotate}()}
}
\concept{coo_modifyers}
\concept{rotations}
