% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{coo_sample}
\alias{coo_sample}
\alias{coo_interpolate}
\title{Sample shapes}
\usage{
coo_sample(x, n, ...)

coo_interpolate(x, n, ...)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{mom_tbl}}

\item{n}{\code{integer} desired number of coordinates (required)}

\item{...}{useless here}
}
\value{
a \link{coo_single}, \link{coo_list} or \link{mom_tbl}
}
\description{
Changes the number of shape coordinates.
}
\details{
\link{coo_sample} will return coordinates regularly sampled along the curvilinear abscissa.
The last point will be dropped so that the distance (new last - first) roughly equals
all other distances between consecutive points. This is typically useful for outlines.
coo_sample_curve (todo link) will preserve the first and last points.
This is typically useful for curves, hence the name.
\link{coo_interpolate} will upsample the number of points.
}
\section{Functions}{
\itemize{
\item \code{coo_interpolate}: Interpolates shape coordinates
}}

\examples{
x <- bot \%>\% pick(1) \%>\% coo_sample(24)
x \%>\% gg()
x \%>\% coo_interpolate(120) \%>\% gg()

bot$coo \%>\% coo_sample(12) \%>\% purrr::map_dbl(nrow)
bot$coo \%>\% coo_sample(12) \%>\% coo_interpolate(24) \%>\% purrr::map_dbl(nrow)

bot \%>\% dplyr::slice(1:5) \%>\% coo_sample(12) -> x
pile(x)
x \%>\% coo_interpolate(48) \%>\% pile()

}
\seealso{
Other coo_modifyers: 
\code{\link{coo_align}()},
\code{\link{coo_center}()},
\code{\link{coo_rotatecenter}()},
\code{\link{coo_rotate}()},
\code{\link{coo_scale}()},
\code{\link{coo_template}()},
\code{\link{coo_trans}()}
}
\concept{coo_modifyers}
\concept{samplers}
