% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{coo_scale}
\alias{coo_scale}
\title{Scale shapes}
\usage{
coo_scale(x, scale)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{coo_tbl}}

\item{scale}{\code{numeric} scaling factor (\link{get_centsize} by default).}
}
\value{
a \link{coo_single}, \link{coo_list} or \link{coo_tbl}
}
\description{
Returns a scaled shape.
}
\examples{

bot2$coo[[1]] \%>\% coo_scale

}
\seealso{
Other coo_modifyers: 
\code{\link{coo_align}()},
\code{\link{coo_center}()},
\code{\link{coo_rotate}()},
\code{\link{coo_sample}()},
\code{\link{coo_template}()},
\code{\link{coo_trans}()}

Other scalings: 
\code{\link{coo_template}()}
}
\concept{coo_modifyers}
\concept{scalings}
