% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{coo_scale}
\alias{coo_scale}
\alias{coo_scale_x}
\alias{coo_scale_y}
\title{Scale shapes}
\usage{
coo_scale(x, scale, from_col, to_col, ...)

coo_scale_x(x, scale, from_col, to_col, ...)

coo_scale_y(x, scale, from_col, to_col, ...)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{mom_tbl}}

\item{scale}{\code{numeric} scaling factor (\link{get_centsize} by default).}

\item{from_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{to_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{...}{useless here}
}
\value{
a \link{coo_single}, \link{coo_list} or \link{mom_tbl}
}
\description{
Returns a scaled shape.
}
\section{Functions}{
\itemize{
\item \code{coo_scale_x}: scale along x-axis

\item \code{coo_scale_y}: scale along y-axis
}}

\examples{

x <- bot \%>\% pick(1)
x \%>\% gg()
x \%>\% coo_scale() \%>\% gg()
x \%>\% coo_scale_x(5) \%>\% gg()
x \%>\% coo_scale_y(5) \%>\% gg()
}
\seealso{
Other coo_modifyers: 
\code{\link{coo_align}()},
\code{\link{coo_baseline}()},
\code{\link{coo_center}()},
\code{\link{coo_reflect}},
\code{\link{coo_rev}()},
\code{\link{coo_rotatecenter}()},
\code{\link{coo_rotate}()},
\code{\link{coo_sample_rr}()},
\code{\link{coo_sample}()},
\code{\link{coo_shear}()},
\code{\link{coo_slide}()},
\code{\link{coo_split}()},
\code{\link{coo_template}()},
\code{\link{coo_trans}()},
\code{\link{coo_trim}()},
\code{\link{coo_up}()}

Other scalings: 
\code{\link{coo_template}()}
}
\concept{coo_modifyers}
\concept{scalings}
