% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom_coo_coe_classes.R
\name{coo_single}
\alias{coo_single}
\alias{new_coo_single}
\alias{validate_coo_single}
\alias{is_coo_single}
\alias{is_coo_single1}
\title{Create single shape made of coordinates}
\usage{
new_coo_single(x = tibble::tibble(x = double(), y = double()))

validate_coo_single(x)

coo_single(x)

is_coo_single(x)

is_coo_single1(x)
}
\arguments{
\item{x}{anything that can be turned a tibble by \code{tibble::tibble()} should work}
}
\value{
a \code{coo_single} object
}
\description{
\code{coo_single} objects are \link[tibble:tibble-package]{tibble} with \emph{exactly} two columns named \code{x} and \code{y}.
Anything that can be turned  should work.
}
\details{
You should use the helper \code{coo_single}, \code{new_coo_single} and \code{validate_coo_single} are for internal operations.
}
\section{Functions}{
\itemize{
\item \code{new_coo_single}: Constructor

\item \code{validate_coo_single}: Validator

\item \code{coo_single}: Helper

\item \code{is_coo_single}: Class tester

\item \code{is_coo_single1}: Class1 tester
}}

\examples{
coo_single()
matrix(1:12, ncol=2) \%>\% coo_single()
list(x=1:2, y=3:4) \%>\% coo_single()
}
