% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{coo_single}
\alias{coo_single}
\alias{is_coo_single}
\alias{is_coo_single1}
\alias{validate_coo_single}
\alias{is_coe_single}
\alias{is_coe_single1}
\alias{validate_coe_single}
\title{coo_single constructor for single shapes}
\usage{
coo_single(x)

is_coo_single(x)

is_coo_single1(x)

validate_coo_single(x)

is_coo_single(x)

is_coo_single1(x)

is_coe_single(x)

is_coe_single1(x)

validate_coe_single(x)
}
\arguments{
\item{x}{anything that can be turned into a coo_single, typically something that looks like a
two columns
@details For \code{tibble} and \code{data.frame}, the first two columns are selected.}
}
\value{
a \code{coo_single} object
}
\description{
\code{coo} objects are \link[tibble:tibble-package]{tibble} with \emph{exactly} two columns named \code{x} and \code{y}
}
\section{Functions}{
\itemize{
\item \code{is_coo_single}: Class tester

\item \code{is_coo_single1}: Class tester

\item \code{validate_coo_single}: Class validator

\item \code{is_coo_single}: Class tester

\item \code{is_coo_single1}: Class tester

\item \code{is_coe_single}: Class tester

\item \code{is_coe_single1}: Class tester

\item \code{validate_coe_single}: Class validator
}}

\examples{
matrix(1:12, ncol=2) \%>\% coo_single()
}
