% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{coo_trans}
\alias{coo_trans}
\title{Translate shapes}
\usage{
coo_trans(x, x_trans = 0, y_trans = 0)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_tbl} or a list of shapes}

\item{x_trans}{\code{numeric} how much translate on x-axis}

\item{y_trans}{\code{numeric} how much translate on y-axis}
}
\value{
\link{coo_single} or \link{coo_tbl} or a list of shapes
}
\description{
Returns a shape translate by \code{x} and \code{y}.
}
\examples{

bot2$coo[[1]] \%>\% coo_trans

}
\seealso{
Other coo_modifyers: 
\code{\link{coo_align}()},
\code{\link{coo_center}()},
\code{\link{coo_scale}()},
\code{\link{coo_template}()}
}
\concept{coo_modifyers}
