% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_modifyers.R
\name{coo_trans}
\alias{coo_trans}
\alias{coo_trans.default}
\alias{coo_trans.coo_list}
\alias{coo_trans.mom_tbl}
\title{Translate shapes}
\usage{
coo_trans(x, x_trans = 0, y_trans = 0, from_col, to_col, ...)

\method{coo_trans}{default}(x, x_trans = 0, y_trans = 0, ...)

\method{coo_trans}{coo_list}(x, x_trans = 0, y_trans = 0, ...)

\method{coo_trans}{mom_tbl}(
  x,
  x_trans = 0,
  y_trans = 0,
  from_col = coo,
  to_col = {     {         from_col     } },
  ...
)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{mom_tbl}}

\item{x_trans, y_trans}{\code{numeric} how much translate on each direction}

\item{from_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{to_col}{colnames from where to get the \link{coo_list}
and how to name the resulting one (only for \link{mom_tbl} method)}

\item{...}{useless here}
}
\value{
a \link{coo_single}, \link{coo_list} or \link{mom_tbl}
}
\description{
Returns a shape translate by \code{x} and \code{y}.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: default method

\item \code{coo_list}: list method

\item \code{mom_tbl}: mom_tbl method
}}

\examples{
bot$coo[[1]] \%>\% coo_trans

}
\seealso{
Other coo_modifyers: 
\code{\link{coo_align}()},
\code{\link{coo_center}()},
\code{\link{coo_rotatecenter}()},
\code{\link{coo_rotate}()},
\code{\link{coo_sample_rr}()},
\code{\link{coo_sample}()},
\code{\link{coo_scale}()},
\code{\link{coo_template}()}

Other translations: 
\code{\link{coo_center}()}
}
\concept{coo_modifyers}
\concept{translations}
