% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_efourier.R
\name{efourier}
\alias{efourier}
\alias{efourier.default}
\alias{efourier.coo_single}
\alias{efourier.coo_tbl}
\alias{efourier_i}
\title{EFT}
\usage{
efourier(x, ...)

\method{efourier}{default}(x, nb_h, ...)

\method{efourier}{coo_single}(x, nb_h, ...)

\method{efourier}{coo_tbl}(x, nb_h, keep_coo = FALSE, ...)

efourier_i(x, nb_pts = 120)
}
\arguments{
\item{x}{plo}

\item{...}{for generics. Useless here.}

\item{nb_h}{\code{int} nb of harmonics}

\item{keep_coo}{\code{logical} whether to retain coo column}

\item{nb_pts}{\code{int} nb of points for the reconstruction}
}
\description{
EFT
}
