% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_descriptors.R
\name{get_centsize}
\alias{get_centsize}
\title{Calculate centroid size}
\usage{
get_centsize(x)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_tbl} or a list of shapes}
}
\value{
\code{numeric} or additional column
}
\description{
Which is the square root of mean squared distances between each point
along the shape and its centroid coordinates.
}
\details{
Can be used, among others, to record size before \link{coo_scale}.
}
\examples{

bot2$coo[[1]] \%>\% get_centsize

}
\seealso{
Other coo_descriptors: 
\code{\link{get_centpos}()}
}
\concept{coo_descriptors}
