% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_getters.R
\name{get_centsize}
\alias{get_centsize}
\alias{get_centsize_norm}
\title{Calculate centroid size}
\usage{
get_centsize(x)

get_centsize_norm(x)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{mom_tbl}}
}
\value{
\code{numeric} or additional column
}
\description{
\link{get_centsize} provides the classical centroid size which is the
square root of the \emph{sum} of squared distances between each point
along the shape and its centroid coordinates. \link{get_centsize_norm} is the
square root of the \emph{mean} of squared distances between each point
along the shape and its centroid coordinates. See Details.
}
\details{
\link{get_centsize_norm} does not depend on the number of points of the shape (see Examples) which is both typical
for outlines and never seen for landmarks.
See Dryden I. E. and Mardia K. V. 1998. Statistical Shape Analysis. Wiley,
Chichester. Since \link{get_centsize} never made a lot of sense to me, as an "outline" person, and because
\link{get_centsize_norm} will "work" in all cases, it is the default method in \link{coo_scale}

Can be used, among others, to record size before \link{coo_scale}.
}
\section{Functions}{
\itemize{
\item \code{get_centsize_norm}: normalized version
}}

\examples{

b <- bot \%>\% pick(1)
b \%>\% get_centsize
b \%>\% coo_sample(64) \%>\% get_centsize()

b \%>\% get_centsize_norm()
b \%>\% coo_sample(64) \%>\% get_centsize_norm()
}
\seealso{
Other coo_descriptors: 
\code{\link{get_centpos}()},
\code{\link{get_range}()}
}
\concept{coo_descriptors}
