% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_getters.R
\name{get_perim_along}
\alias{get_perim_along}
\alias{get_perim}
\alias{get_perim_cumsum}
\title{Calculates perimeter and variations}
\usage{
get_perim_along(x)

get_perim(x)

get_perim_cumsum(x)
}
\arguments{
\item{x}{\link{coo_single}, \link{coo_list} or \link{coo_tbl}}
}
\value{
\code{numeric} or additional column
}
\description{
\itemize{
\item \code{get_perim_along} calculates the euclidean distance between every points of a shape
\item \code{get_perim} is simply \code{sum(coo_perim_along)}
\item \code{get_perim_cumsum} is simply \code{cumsum(coo_perim_along)}
}
}
\section{Functions}{
\itemize{
\item \code{get_perim}: Calculates total perimeter

\item \code{get_perim_cumsum}: Calculates cumsum between successive points of a shape
}}

\examples{
bot2 \%>\% pick(1) \%>\% get_perim_along()
}
\concept{getters}
\concept{perimeter getters}
