% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_graphics.R, R/stat_pca.R
\name{gg0}
\alias{gg0}
\alias{gg0.default}
\alias{gg0.coo_single}
\alias{gg0.coo_tbl}
\alias{gg0.tbl}
\alias{gg}
\alias{gg0.pca}
\title{Universal graphics functions}
\usage{
gg0(x, ...)

\method{gg0}{default}(x, ...)

\method{gg0}{coo_single}(x, ...)

\method{gg0}{coo_tbl}(x, ...)

\method{gg0}{tbl}(x, ...)

gg(x, first = TRUE, centroid = TRUE, axes = TRUE, ...)

\method{gg0}{pca}(x, x_axis, y_axis, ...)
}
\arguments{
\item{x}{a Momocs object}

\item{...}{additional parameters to feed geoms}

\item{first}{\code{logical} whether to draw first point}

\item{centroid}{\code{logical} whether to draw centroid}

\item{axes}{\code{logical} whether to draw axes, text and grid}

\item{x_axis}{column name}

\item{y_axis}{column name}

\item{gg}{\code{ggplot} object, default to \link[ggplot2:last_plot]{ggplot2::last_plot}}
}
\value{
a \code{ggplot} object
}
\description{
Default (ggplot2) visualisations for Momocs objects.
}
\details{
\code{gg} is the base plotter.
\code{gg0} prepare the canvas but let you pick your \code{ggplot2::geoms}.
\code{draw} add shapes on top of last plot
}
\note{
I call it "universal" as a reminder to provide a gg for each object.
}
\examples{
b <- bot2 \%>\% pick(5)
gg(b)

# Let's add some geoms to
nice_plot <- gg0(b) +
    ggplot2::geom_polygon(color="grey50", fill="red", alpha=0.25) +
    ggplot2::geom_point(shape="circle plus")
nice_plot # print it

# you have all ggplot2 for free 8-)
# you do not have to ggplot2:: if you library(ggplot2) before
gorgeous_plot <- nice_plot + ggplot2::theme_minimal() +
    ggplot2::labs(x="abscissa", y="ordinate", title="Drink responsibly")

# this is a plotting factory !
# gorgeous_plot \%+\% pick(bot2, 12)
}
