% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_graphics.R
\name{gg0}
\alias{gg0}
\alias{gg}
\alias{gg.default}
\alias{gg.coo_single}
\title{ggplot2 default plotter}
\usage{
gg0(x, ...)

gg(x, ...)

\method{gg}{default}(x, ...)

\method{gg}{coo_single}(x, first = TRUE, centroid = TRUE, axes = TRUE, ...)
}
\arguments{
\item{x}{a Momocs object}

\item{...}{additional parameters to feed geoms}

\item{first}{\code{logical} whether to draw first point}

\item{centroid}{\code{logical} whether to draw centroid}

\item{axes}{\code{logical} whether to draw axes, text and grid}
}
\value{
a \code{ggplot} object
}
\description{
Default ((gplot2) visualisations for Momocs objects.
}
\details{
\code{gg0} prepare the canvas but let you pick your \code{ggplot2::geoms}.
}
\examples{
bot2 \%>\% pick(1) \%>\% gg()
bot2 \%>\% pick(1) \%>\% gg0() + ggplot2::geom_point(shape="circle plus")

}
