% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hearts}
\alias{hearts}
\title{Heart dataset}
\format{
A data frame with 40 rows and three columns
\describe{
\item{img}{dummy paths}
\item{coo}{outline coordinates}
\item{coo_ldk}{landmark id (4) on these coordinates}
\item{author}{who drawn these}
}
}
\source{
We thank the fellows of the Ecology Department of the
French Institute of Pondicherry circa 2011 that drawn the hearts.
They then have been landmarked, smoothed, scaled, centered, and
downsampled to 80 coordinates per outline.
}
\usage{
hearts
}
\description{
A list of 4 shapes (as outlines)
}
\keyword{datasets}
