% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cool_verbs.R
\name{list_columns}
\alias{list_columns}
\alias{coo_nb}
\alias{coe_nb}
\alias{coo_only}
\alias{coe_only}
\alias{coo_drop}
\alias{coe_drop}
\alias{coo_select}
\alias{coe_select}
\alias{coo_names}
\alias{coe_names}
\title{List columns verbs}
\usage{
coo_nb(x)

coe_nb(x)

coo_only(x)

coe_only(x)

coo_drop(x)

coe_drop(x)

coo_select(x, ...)

coe_select(x, ...)

coo_names(x)

coe_names(x)
}
\arguments{
\item{x}{\link{coo_tbl} or a \link{coe_tbl}}

\item{...}{additional parameters, just like in \code{dplyr::select}}
}
\description{
List columns verbs
}
\section{Functions}{
\itemize{
\item \code{coo_nb}: count the number of coo columns

\item \code{coe_nb}: count the number of coe columns

\item \code{coo_only}: only retain coo columns

\item \code{coe_only}: only retain coe columns

\item \code{coo_drop}: drop coo columns

\item \code{coe_drop}: drop coe columns

\item \code{coo_select}: use select among coo columns

\item \code{coe_select}: use select among coe columns

\item \code{coo_names}: return names of coo columns

\item \code{coe_names}: return names of coe columns
}}

