% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momocs2-package.R
\docType{package}
\name{Momocs2-package}
\alias{Momocs2}
\alias{Momocs2-package}
\title{Momocs2: Morphometrics methods for outlines, curves, landmarks and measurements}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The goal of 'Momocs' is to provide a complete, convenient, reproducible and open-source toolkit for 2D morphometrics. It includes most common 2D morphometrics approaches on outlines, open outlines, configurations of landmarks, traditional morphometrics, and facilities for data preparation, manipulation and visualization with a consistent grammar throughout. It allows reproducible, complex morphometrics analyses and other morphometrics approaches should be easy to plug in, or develop from, on top of this canvas. Momocs2 is part of MomX and a complete rewrite of Momocs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MomX/Momocs2}
  \item \url{https://momx.github.io/Momocs2/}
  \item Report bugs at \url{https://github.com/MomX/Momocs2/issues}
}

}
\author{
\strong{Maintainer}: Vincent Bonhomme \email{bonhomme.vincent@gmail.com}

Authors:
\itemize{
  \item Julien Claude (core functions in base R)
}

}
\keyword{internal}
