% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{olea}
\alias{olea}
\title{Olea dataset}
\format{
A data frame with 90 rows (three accessions):
\describe{
\item{id}{unique id}
\item{VD}{curves coordinates for dorsal view}
\item{VL}{curves coordinates for lateral view}
\item{var}{olive accession name}
\item{status}{factor, whether accession is wild or cultivated}
}
}
\source{
We thank Jean-Frederic Terral and Sarah Ivorra
(UMR CBAE, Montpellier, France) from allowing us to share the data.

You can have a look to the original paper:
Terral J-F, Alonso N, Capdevila RB i, Chatti N, Fabre L, Fiorentino G,
Marinval P, Jorda GP, Pradat B, Rovira N, et al. 2004.
Historical biogeography of olive domestication (\emph{Olea europaea} L.)
as revealed by geometrical morphometry applied to biological and archaeological material.
\emph{Journal of Biogeography} 31: 63-77.
}
\usage{
olea
}
\description{
An open curves dataset with two views derived from the legacy Momocs olea dataset.
}
\details{
Olive stones, photographed using two orthogonal views.
}
\keyword{datasets}
