% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom_coo_coe_verbs.R
\name{pick}
\alias{pick}
\title{Pick a single coo_df from a mom_tbl}
\usage{
pick(x, i)
}
\arguments{
\item{x}{a \code{mom_tbl} or a \code{coo_list}}

\item{i}{\code{int} which to extract. If not provided, pick one randomly.}
}
\value{
a coo_df
}
\description{
Pick a single coo_df from a mom_tbl
}
\note{
Used to be \code{[[} in Momocs < 2.0 (eg \code{bot[[1]]} is replace with \code{bot \%>\% pick(1)})
}
\examples{
bot \%>\% pick(1) \%>\% gg()
bot$coo \%>\% pick(2)
}
\seealso{
\link{slive}

Other mom verbs: 
\code{\link{press}()},
\code{\link{slive}()},
\code{\link{unfold}()}
}
\concept{mom verbs}
