% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momocs_verbs.R
\name{pick}
\alias{pick}
\title{Pick a single coo_df}
\usage{
pick(x, i)
}
\arguments{
\item{x}{a \code{coo_tbl} or a \code{coo_list}}

\item{i}{\code{int} which to extract. If not provided, pick one randomly.}
}
\value{
a coo_df
}
\description{
Pick a single coo_df
}
\examples{
bot2 \%>\% pick(1) \%>\% gg()
bot2$coo \%>\% pick(2)
}
