% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_graphics.R
\name{pile}
\alias{pile}
\alias{pile0}
\title{Plot all shapes on the same graph}
\usage{
pile0(x)

pile(x, ..., from_col, geom, size, alpha)
}
\arguments{
\item{x}{a mom_tbl object}

\item{...}{additional parameters to pass to main geom}

\item{from_col}{column name to use for plotting (only for \code{mom_tbl} where it defaults to \code{coo})}

\item{geom}{ggplot2 geom to use, typically \code{geom_path} or \code{geom_point}}

\item{size, alpha}{to feed geom}
}
\description{
Plot all shapes on the same graph
}
\section{Functions}{
\itemize{
\item \code{pile0}: prepare

\item \code{pile}: plotter
}}

\note{
formerly named \code{stack} in retired Momocs.
\code{stack} was quite a fine name but a \code{stack} is not something more fundamental in informatics.
Also, \code{pile} is shorter.
}
\examples{
bot \%>\% pick(1) \%>\% pile()
pile(bot$coo)
bot \%>\% dplyr::rename(coo2=coo) \%>\% pile(from_col=coo2)
}
\concept{coo_graphics}
