% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_graphics.R
\name{pile}
\alias{pile}
\title{Plot all shapes on the same graph}
\usage{
pile(x, ...)
}
\arguments{
\item{x}{a coo_tbl object}

\item{...}{additional parameters to feed ggplot2::geoms}
}
\description{
Plot all shapes on the same graph
}
\details{
\code{pile_column} is an helper function that takes a \code{coo_list} and
returns a single \code{tbl} with an additional name/index column.
}
\note{
former stack in old Momocs
}
\examples{
bot2 \%>\% pile()
# global spec
bot2 \%>\% pile(col="gold")
# aes bounded to x columns
bot2 \%>\% pile(type) # you can omit ggplot2:: if it's loaded

# a more complex example
bot2 \%>\% pile(type) +
   ggplot2::facet_grid(~type) +
   ggplot2::scale_colour_manual(values=c("forestgreen", "orange"))
}
\seealso{
Other family_picture: 
\code{\link{inspect}()},
\code{\link{mosaic}()}
}
\concept{family_picture}
