% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_printers.R
\name{pillar_methods}
\alias{pillar_methods}
\alias{type_sum.coo}
\alias{type_sum.out}
\alias{type_sum.ldk}
\alias{type_sum.ldk_id}
\alias{type_sum.path}
\alias{type_sum.cur}
\alias{type_sum.dct}
\alias{type_sum.eft}
\alias{type_sum.npoly}
\alias{type_sum.opoly}
\alias{type_sum.proc}
\alias{pillar_shaft.coo}
\alias{pillar_shaft.out}
\alias{pillar_shaft.ldk}
\alias{pillar_shaft.ldk_id}
\alias{pillar_shaft.path}
\alias{pillar_shaft.cur}
\alias{pillar_shaft.dct}
\alias{pillar_shaft.eft}
\alias{pillar_shaft.npoly}
\alias{pillar_shaft.opoly}
\alias{pillar_shaft.proc}
\title{Pillar support for morphometric data types}
\usage{
type_sum.coo(x)

type_sum.out(x)

type_sum.ldk(x)

type_sum.ldk_id(x)

type_sum.path(x)

type_sum.cur(x)

type_sum.dct(x)

type_sum.eft(x)

type_sum.npoly(x)

type_sum.opoly(x)

type_sum.proc(x)

\method{pillar_shaft}{coo}(x, ...)

\method{pillar_shaft}{out}(x, ...)

\method{pillar_shaft}{ldk}(x, ...)

\method{pillar_shaft}{ldk_id}(x, ...)

\method{pillar_shaft}{path}(x, ...)

\method{pillar_shaft}{cur}(x, ...)

\method{pillar_shaft}{dct}(x, ...)

\method{pillar_shaft}{eft}(x, ...)

\method{pillar_shaft}{npoly}(x, ...)

\method{pillar_shaft}{opoly}(x, ...)

\method{pillar_shaft}{proc}(x, ...)
}
\description{
These functions provide custom display methods for morphometric classes
in tibble printing via the pillar package.
}
\keyword{internal}
