% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plint.R
\name{plint}
\alias{plint}
\title{Print, plot and continue the pipe}
\usage{
plint(x, ...)
}
\arguments{
\item{x}{object to plint}

\item{...}{additional parameters}
}
\description{
Print, plot and continue the pipe
}
