% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momocs_verbs.R
\name{unfold}
\alias{unfold}
\title{Unfold list columns}
\usage{
unfold(x, ...)
}
\arguments{
\item{x}{a Momocs object}

\item{...}{useless}
}
\description{
Unfold list columns such as \link{coo_list} and \link{coe_list}, repeat lines if required.
}
\note{
Differs from \code{tidyr::unnest}, in that it adds a "shp" column.
Also \code{unpack} was already take by \code{tidyr}. \code{unfold.list} is just \code{dplyr::bind_rows}.
}
\examples{

bot2$coo \%>\% unfold

bot2 \%>\% unfold
}
