% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momocs_verbs.R
\name{unpack}
\alias{unpack}
\title{Unpack list columns}
\usage{
unpack(x, ...)
}
\arguments{
\item{x}{a Momocs object}

\item{...}{useless}
}
\description{
Unpack list columns, repeat lines if required. Differs from unnest,
in that it adds a "shp" column.
}
\examples{

bot2$coo \%>\% unpack

bot2 \%>\% unpack
}
